Add files via upload
I added the comboBox and set it up. Also I found a solution to our issue writing our global data to the TanFile object. I made a TanFile object named project. When writing information to the project, simply use this object to store data. This object can now be accessed by anything in the MainWindow class.