Connecting Cells click() to Signals/Slots Dynamically
Utilizing findChild() to obtain references to each cell.
Still having to create each cell name by hand.
Still uncertain how to check that the cell object is valid, since
findChild returns a 0 if child not found. How do I check for that, when
the stupid thing won't let me see if an object pointer == 0? Thus, the
program will not compile if the object names are not found correctly.