Renovating Frame Display Using Custom Classes
Replaced QPushButtons with custom CellWidget class.
Rewrote all relevant functions that create the frame.
Created functions that should talk between TanFile data representation
and the individual cells of the display frame.

Compiles, but still need to test -- waiting for TanFile_Save.cpp to be
corrected so it compiles.
Still need to finish implementing the mouse event functions and the
function that updates the TanFile array when color changes in cell
widget -- waiting on update to TanFile, private to public.